+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implement logout for users accounts [#329](https://github.com/rokwire/core-building-block/issues/329)
 
 ## [1.20.1] - 2022-03-07
 ### Fixed